--- http/sajjad/homework_1_http/go.mod
@@ -1,3 +1,0 @@
-module homework_1_http
-
-go 1.24.1
