--- http/sajjad/homework_1_http/models/user.go
@@ -1,11 +1,0 @@
-package models
-
-type User struct {
-	Attributes Attributes `json:"attributes"`
-}
-
-type Attributes struct {
-	Id         int    `json:"id"`
-	Name       string `json:"name"`
-	AccountIds []int  `json:"account_ids"`
-}
