+++ http/sajjad/http-client/models/account.go
@@ -1,12 +1,14 @@
 package models
 
+// Account represents a user account with attributes
 type Account struct {
 	Attributes AccountAttributes `json:"attributes"`
 }
 
+// AccountAttributes contains the account's attributes
 type AccountAttributes struct {
-	Id      int    `json:"id"`
-	UserId  int    `json:"user_id"`
+	ID      int    `json:"id"`
+	UserID  int    `json:"user_id"`
 	Name    string `json:"name"`
 	Balance int    `json:"balance"`
 }
